+++ src/tests/Warp.Tests/Adapters/AdapterEndpointTestsBase.cs
@@ -13,8 +13,8 @@ using Warp.Core.Data.Entities;
 using Warp.Core.Enums;
 using Warp.Core.Services;
 using Warp.Tests.Fixtures;
+using Warp.UI;
 using Warp.UI.Endpoints;
-using Warp.UI.UIMiddleware;
 
 namespace Warp.Tests.Adapters;